fix: ci报错问题
