+++ package.json
@@ -107,7 +107,7 @@
   "limit-size": [
     {
       "path": "dist/larkmap.min.js",
-      "limit": "700 Kb"
+      "limit": "750 Kb"
     }
   ],
   "keywords": [
